# ParkingBuddy/OpenAPIs/openapi.json
{"openapi":"3.0.1","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/new":{"post":{"summary":"Create New User","description":"Creates a new user\n\nArgs:\n    create_user (CreateUserModel): Data\n\nReturns:\n    JSONResponse: Response","operationId":"create_new_user_users_new_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"201":{"description":"Created new user"},"400":{"description":"Invalid data provided for user"},"409":{"description":"User with provided data already exists"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/data":{"post":{"summary":"Get User Data","description":"Gets user data\n\nArgs:\n    user (UserModel): Input data\n\nReturns:\n    UserDataModel: User data","operationId":"get_user_data_users_data_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}},"required":true},"responses":{"200":{"description":"Got user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataModel"}}}},"400":{"description":"Invalid data provided for user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/plates/new":{"post":{"summary":"Create New Plate","description":"Creates a new registration plate\n\nArgs:\n    create_plate (CreatePlateModel): Data\n\nReturns:\n    JSONResponse: Response","operationId":"create_new_plate_plates_new_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlateModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"201":{"description":"Created new registration plate"},"400":{"description":"Invalid data provided"},"409":{"description":"Plate with provided data already exists"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entry/in":{"post":{"summary":"Create New Entry","description":"Creates a new parking entry\n\nArgs:\n    create_entry (CreateEntryModel): Data\n\nReturns:\n    JSONResponse: Response","operationId":"create_new_entry_entry_in_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEntryModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"201":{"description":"Created new parking entry"},"400":{"description":"Invalid data"},"409":{"description":"Plate with provided data already parked"},"402":{"description":"Previous parking has not been paid. Vehilcle cannot enter"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entry/out":{"post":{"summary":"Entry Exit","description":"Exit entered car\n\nArgs:\n    exit_entry (ExitEntryModel): Data\n\nReturns:\n    JSONResponse: Response","operationId":"entry_exit_entry_out_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExitEntryModel"}}},"required":true},"responses":{"200":{"description":"Exited successfully","content":{"application/json":{"schema":{}}}},"400":{"description":"Invalid data"},"409":{"description":"Plate with provided data never parked"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entry/pay":{"post":{"summary":"Pay Entry","description":"Pay exited car to allow next entry\n\nArgs:\n    pay_entry (PayEntryModel): Data\n\nReturns:\n    JSONResponse: Response","operationId":"pay_entry_entry_pay_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayEntryModel"}}},"required":true},"responses":{"200":{"description":"Paid successfully","content":{"application/json":{"schema":{}}}},"202":{"description":"Entry already paid"},"400":{"description":"Invalid data"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entry/data":{"post":{"summary":"Get Entry Data","description":"Gets entry data\n\nArgs:\n    registration (RegistrationModel): Input data\n\nReturns:\n    list[EntryDataModel]: Entry data","operationId":"get_entry_data_entry_data_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationModel"}}},"required":true},"responses":{"200":{"description":"Got entry data","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/EntryDataModel"},"type":"array","title":"Response Get Entry Data Entry Data Post"}}}},"400":{"description":"Invalid data provided for entry"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CreateEntryModel":{"properties":{"registration_plate":{"type":"string","title":"Registration Plate"}},"type":"object","required":["registration_plate"],"title":"CreateEntryModel"},"CreatePlateModel":{"properties":{"email":{"type":"string","title":"Email"},"registration_plate":{"type":"string","title":"Registration Plate"}},"type":"object","required":["email","registration_plate"],"title":"CreatePlateModel"},"CreateUserModel":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"email":{"type":"string","title":"Email"}},"type":"object","required":["username","password","email"],"title":"CreateUserModel"},"EntryDataModel":{"properties":{"id":{"type":"integer","title":"Id","default":0},"entry_time":{"type":"string","format":"date-time","title":"Entry Time","default":"2024-05-29T17:02:26.368782"},"exit_time":{"type":"string","format":"date-time","title":"Exit Time","default":"2024-05-29T17:02:26.368789"},"is_paid":{"type":"boolean","title":"Is Paid","default":false},"registration_plate":{"type":"string","title":"Registration Plate","default":""}},"type":"object","title":"EntryDataModel"},"ExitEntryModel":{"properties":{"registration_plate":{"type":"string","title":"Registration Plate"}},"type":"object","required":["registration_plate"],"title":"ExitEntryModel"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PayEntryModel":{"properties":{"entry_id":{"type":"integer","title":"Entry Id"}},"type":"object","required":["entry_id"],"title":"PayEntryModel"},"RegistrationModel":{"properties":{"registration_plate":{"type":"string","title":"Registration Plate"}},"type":"object","required":["registration_plate"],"title":"RegistrationModel"},"UserDataModel":{"properties":{"username":{"type":"string","title":"Username","default":""},"email":{"type":"string","title":"Email","default":""},"registration_plates":{"items":{"type":"string"},"type":"array","title":"Registration Plates","default":[]}},"type":"object","title":"UserDataModel"},"UserModel":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserModel"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}